+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [1.26.0](https://github.com/posit-dev/team-operator/compare/v1.25.0...v1.26.0) (2026-05-22)
+
+
+### Features
+
+* pin Positron Pro version via positronConfig.version ([#138](https://github.com/posit-dev/team-operator/issues/138)) ([06d9433](https://github.com/posit-dev/team-operator/commit/06d94334798c971484625f30230e7de818693193))
+
+# [1.25.0](https://github.com/posit-dev/team-operator/compare/v1.24.1...v1.25.0) (2026-05-12)
+
+
+### Features
+
+* add livenessProbe to Package Manager ([#137](https://github.com/posit-dev/team-operator/issues/137)) ([ac44f4c](https://github.com/posit-dev/team-operator/commit/ac44f4c9f2e54296ab5da844c3ecfa9a4ed7c920))
+
 ## [1.24.1](https://github.com/posit-dev/team-operator/compare/v1.24.0...v1.24.1) (2026-05-07)